added truncation script, improved organisation

+++ SolubilityCheck.py
@@ -2,13 +2,16 @@ import os
 import parse_results
 import glob
 
-blast_dir = "/Users/gera/Downloads/blast-2.2.18/bin"
-folder = "/Users/gera/PycharmProjects/SolubilityCheck/genomes/"
-all_genomes=glob.glob(folder+"/*.fa.gz")
-#db = "/Users/gera/PycharmProjects/SolubilityCheck/genomes/Arabidopsis_halleri.Ahal2.2.dna.toplevel.fa.gz"
-db_zipped = ""
-query = blast_dir+"/RNH.fa"
 
+#  this scrip uses protein query to run blastp against genomes of interest and
+#  calculates the solubility of the resulting sequences
+
+blast_dir = "/Users/gera/Downloads/blast-2.2.18/bin"  # your blast bin folder
+folder = "/Users/gera/PycharmProjects/SolubilityCheck/genomes/"  # here you store archived genomes (in *.fa.gz format)
+all_genomes = glob.glob(folder+"/*.fa.gz")
+query = blast_dir+"/RNH.fa"  # here is your  query file (somehow should be in the blast bin folder)
+
+db_zipped = ""
 for db in all_genomes:
     # checking the format of the genome, we'd like to have them all zipped after processing
     if ".gz" in db:
@@ -30,29 +33,32 @@ for db in all_genomes:
             print("gunzip " + db_zipped)
             os.system("gunzip " + db_zipped)
         request1 = "makeblastdb -in "+db+" -dbtype nucl"
-        request2 = "tblastn -query "+query+" -db "+db+" -out "+out+" -qcov_hsp_perc 10 -outfmt \"6 sseqid sstart send sseq\""
+        request2 = "tblastn -query "+query+" -db "+db+" -out "+out+" -outfmt \"6 sseqid sstart send sseq\""
         print(request1)
         print(request2)
         os.system(request1)
         os.system(request2)
 
     # processing results of tblastn
-    my_file = open(out, "r")
-    seq_file = open(seq, "w")
-    line2 = my_file.readline()
-    count = 0
-    all_hits = []
-    while line2:
-        if "*" not in line2:
-            line2 = line2.rstrip()
-            my_list = (line2.split("\t"))
-            my_seq = my_list.pop(-1)
-            if len(my_seq) > 500:
-                joined_string = "|".join([str(v) for v in my_list])
-                print(">", joined_string, file=seq_file, sep="")
-                print(my_seq, file=seq_file)
+    if os.path.exists(seq) and os.path.getsize(seq) > 0:
+        print("Blast results are already parsed, skipping")
+    else:
+        my_file = open(out, "r")
+        seq_file = open(seq, "w")
         line2 = my_file.readline()
-    seq_file.close()
+        count = 0
+        all_hits = []
+        while line2:
+            if "*" not in line2:
+                line2 = line2.rstrip()
+                my_list = (line2.split("\t"))
+                my_seq = my_list.pop(-1)
+                if len(my_seq) > 500:
+                    joined_string = "|".join([str(v) for v in my_list])
+                    print(">", joined_string, file=seq_file, sep="")
+                    print(my_seq, file=seq_file)
+            line2 = my_file.readline()
+        seq_file.close()
 
     # running Protein-Sol and parsing results
     file2 = db+"_seq_prediction.txt"
@@ -67,7 +73,7 @@ for db in all_genomes:
             request5 = "cp seq_prediction.txt "+file2
             print(request5)
             os.system(request5)
-            parse_results.parse_solubility_results(file2, genome_name)
+            parse_results.parse_solubility_results(file2, genome_name, True, 'bar')
         else:
             print("No homologues of"+query+"were found in"+genome_name)
 
@@ -81,3 +87,5 @@ for db in all_genomes:
     else:
         print("zipping " + db)
         os.system("gzip " + db)
+
+    print("++++++++++DONE++++++++++")
